backoffice: generalize views  * ref https://github.com/cern-sis/issues-inspire/issues/960
